+++ lib/src/languages/dart/dart_grammar.dart
@@ -633,6 +633,7 @@ class DartGrammarDefinition extends DartGrammarLexer {
               expressionVariableDirectOperation() |
               expressionVariableAssigment() |
               expressionFunctionInvocation() |
+              expressionObjectEntryFunctionInvocation() |
               expressionVariableEntryAccess() |
               expressionGetterAccess() |
               expressionNullValue() |
@@ -743,6 +744,31 @@ class DartGrammarDefinition extends DartGrammarLexer {
         var expression = v[2];
         return ASTExpressionVariableEntryAccess(variable, expression);
       });
+
+  Parser<ASTExpressionObjectEntryFunctionInvocation>
+  expressionObjectEntryFunctionInvocation() =>
+      (variable() &
+              char('[') &
+              ref0(expression) &
+              char(']') &
+              char('.').trimHidden() &
+              identifier() &
+              char('(').trimHidden() &
+              ref0(expressionSequence).optional() &
+              char(')').trimHidden())
+          .map((v) {
+            var variable = v[0];
+            var expression = v[2];
+            var fName = v[5];
+            var args = v[7];
+            args ??= <ASTExpression>[];
+            return ASTExpressionObjectEntryFunctionInvocation(
+              variable,
+              expression,
+              fName,
+              args,
+            );
+          });
 
   Parser<ASTExpressionListLiteral> expressionListEmptyLiteral() =>
       ((char('<').trimHidden() & simpleType() & char('>').trimHidden())